sdm660-common: rootdir: Optimize vfs cache pressure
Xiaomi SDM660 have 3/4GB of RAM and if we keep high value then it will
makes the RAM too busy which will hurt the performance and if the value
is below 20 then it will make the device slow after few days as most of
the cache are unused.

Let's keep it '40' - neither too high nor too low.

Change-Id: I5787f38d2a5f43c7dc4f1dc256173cf700052875
Signed-off-by: ImSpiDy <SpiDy2713@gmail.com>